\name{DAILY.QA.DIRECTORY}
\alias{DAILY.QA.DIRECTORY}
\docType{data}
\title{ Data after QA flags applied
 
}
\description{ Data is downloaded with various QA flags. After they have
   have been applied the  data is given a .dat  extention and placed in the QA
    directory
 
}
\usage{ DAILY.QA.DIRECTORY }
\format{
  The format is:
 chr "DailyData"
}
\details{
  see ftp readme for flags
}
\source{
 
}
\references{
 
}
\examples{
print(DAILY.QA.DIRECTORY)
 
}
\keyword{directories}
