\name{GHCN.DAILY.COUNTRY.URL}
\alias{GHCN.DAILY.COUNTRY.URL}
\docType{data}
\title{Url of the country FIPS data
 
}
\description{Url to the country FIPS codes
 
}
\usage{ GHCN.DAILY.COUNTRY.URL }
\format{
  The format is:
 chr "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/ghcnd-countries.txt"
}
\details{ current URL to the NCDC file of FIPS codes
 
}
\source{\url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/ghcnd-countries.txt}
 
}
\references{\url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/ghcnd-countries.txt}
   
}
\examples{
print(GHCN.DAILY.COUNTRY.URL)
 
}
\keyword{datasets}
