\name{GHCN.DAILY.INVENTORY.URL}
\alias{GHCN.DAILY.INVENTORY.URL}
\docType{data}
\title{ url for inventory
 
}
\description{ the url for the file of station inventories
 
}
\usage{ GHCN.DAILY.INVENTORY.URL }
\format{
  The format is:
 chr "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/ghcnd-inventory.txt"
}
\details{ This url is used for downloading the file
 
}
\source{The source is NCDC see the readme below
 
}
\references{\url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt}
 
}
\examples{
print(GHCN.DAILY.INVENTORY.URL)
 
}
\keyword{datasets}
