\name{GHCN.DAILY.METADATA.URL}
\alias{GHCN.DAILY.METADATA.URL}
\docType{data}
\title{ url to station metadata
 
}
\description{ url to station metadata on the NCDCftp site.
  This file contains station location, name and identifier
  information
 
}
\usage{ GHCN.DAILY.METADATA.URL }
\format{
  The format is:
 chr "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/ghcnd-stations.txt"
}
 
\source{ NCDC is the source
 
}
\references{ url{ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt}
 
}
\examples{
print(GHCN.DAILY.METADATA.URL)
 
}
\keyword{datasets}
