\name{MONTHLY.DATA.DIRECTORY}
\alias{MONTHLY.DATA.DIRECTORY}
\docType{data}
\title{A directory for monthly data
 
}
\description{Monthly data is complied into datasets that contain all
  stations. Both \code{writeNetCDF} and \code{writeGhcn} write
  data to this directory
 
}
\usage{ MONTHLY.DATA.DIRECTORY }
\format{
  The format is:
 chr "MonthlyData"
}
\details{ A directory for Monthly datasets
 
}
 
 
\examples{
print(MONTHLY.DATA.DIRECTORY)
 
}
\keyword{directories}
