\name{readCountry}
\alias{readCountry}
 
\title{read country Fips codes
 
}
\description{reads the FIPS codes
 
}
\usage{
readCountry(filename = "countryCodes.txt")
}
 
\arguments{
  \item{filename}{Name of the file
 
}
}
 
\value{Returns a dataframe
 
}
\references{
 insert
}
\author{Steven Mosher
 
}
 

  
\examples{
 \dontrun{
 #sample
 }
}
 
\keyword{ filetools }
 
