\name{DAILY.DATA.DIRECTORY}
\alias{DAILY.DATA.DIRECTORY}
\docType{data}
\title{Directory for station data
 
}
\description{The local directory where all functions expect to find
  the source data. All files here have a .dat extension
 
}
\usage{ DAILY.DATA.DIRECTORY }
\format{
  The format is:
 chr "DailyData"
}
\details{ The data directory name. Avoid changing it unless understand the
  code completely. Recall that on start this value will always be read
  into the variable.  
 
}
 
 
\examples{
print(DAILY.DATA.DIRECTORY)
 
}
\keyword{datasets}
